+++ src/App/Header/MobileNav/Menu.tsx
@@ -3,8 +3,8 @@ import ThemeToggle from "App/Header/ThemeToggle";
 import { NavLink } from "react-router-dom";
 import { Link } from "../../types";
 import { useModalContext } from "contexts/ModalContext";
-import APLogo, { LogoProps } from "components/APLogo";
 import Icon from "components/Icon";
+import Logo, { LogoProps } from "components/Logo";
 import { createNavLinkStyler } from "helpers";
 
 export default function Menu({
@@ -21,7 +21,7 @@ export default function Menu({
       className="fixed top-0 inset-x-0 z-10 bg-blue dark:bg-blue-d5 shadow-lg pb-8"
     >
       <div className="flex justify-between items-center w-full py-4 padded-container border-b border-gray-l2">
-        <APLogo className="w-32" logo={logo} />
+        <Logo className="w-32" logo={logo} />
         <button
           onClick={closeModal}
           className="flex items-center text-white justify-center"
