+++ src/App/Header/index.tsx
@@ -2,7 +2,7 @@ import Airdrop from "App/Header/Airdrop";
 import { useEffect, useRef, useState } from "react";
 import { Location, matchRoutes, useLocation } from "react-router-dom";
 import { Link } from "../types";
-import APLogo, { LogoProps } from "components/APLogo";
+import Logo, { LogoProps } from "components/Logo";
 import WalletSuite from "components/WalletSuite";
 import { appRoutes } from "constants/routes";
 import DesktopNav from "./DesktopNav";
@@ -41,7 +41,7 @@ export default function Header({ classes, links, logo }: Props) {
       transition-shadow ease-in-out duration-300 w-full h-[90px]`}
     >
       <div className="grid items-center gap-4 padded-container grid-cols-[auto_1fr_auto] h-full">
-        <APLogo className="w-32" logo={logo} />
+        <Logo className="w-32" logo={logo} />
         <DesktopNav classes="hidden lg:flex" links={links} />
         <div className="flex gap-4 justify-self-end">
           <ThemeToggle classes="hidden lg:flex" />
