+++ src/App/Layout.tsx
@@ -2,8 +2,8 @@ import { Suspense } from "react";
 import { Outlet } from "react-router-dom";
 import { Link } from "./types";
 import angelProtocolLogo from "assets/images/angelprotocol-beta-horiz-wht.svg";
-import { LogoProps } from "components/APLogo";
 import Loader from "components/Loader";
+import { LogoProps } from "components/Logo";
 import Seo from "components/Seo";
 import { appRoutes } from "constants/routes";
 import Footer from "./Footer";
