+++ src/components/Logo.tsx
@@ -8,7 +8,7 @@ export type LogoProps = {
 
 type Props = { className?: string; logo: LogoProps };
 
-export default function APLogo({ className, logo }: Props) {
+export default function Logo({ className, logo }: Props) {
   return (
     <ExtLink href={logo.href} title={logo.title} className={className}>
       <img src={logo.src} alt="" className="w-full h-full" />
