+++ src/pages/Admin/Charity/WidgetConfigurer/WidgetExample/WidgetExample.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from "react";
 import { useFormContext } from "react-hook-form";
 import angelProtocolLogo from "assets/images/angelprotocol-beta-horiz-wht.svg";
-import APLogo, { LogoProps } from "components/APLogo";
+import Logo, { LogoProps } from "components/Logo";
 import { getPossessiveForm } from "helpers";
 import { FormValues } from "../schema";
 import Donater from "./Donater";
@@ -70,7 +70,7 @@ export default function WidgetExample({ trigger }: { trigger: boolean }) {
           </section>
         </div>
         <footer className="flex justify-center items-center h-20 w-full bg-blue">
-          <APLogo className="w-20" logo={LOGO} />
+          <Logo className="w-20" logo={LOGO} />
         </footer>
       </div>
     </div>
